LayoutTests/sputnik/Conformance/13_Function_Definition/S13_A7_T3-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unrecognized token '\'
S13_A7_T3

PASS Expected parsing failure

TEST COMPLETE
